+++ assembly/pom.xml
@@ -26,6 +26,11 @@
             <version>${project.parent.version}</version>
         </dependency>
         <dependency>
+            <groupId>com.pingcap.tispark</groupId>
+            <artifactId>spark-wrapper-spark-2.3</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
+        <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
             <version>${mysql.connector.version}</version>
