+++ core/src/main/scala/org/apache/spark/sql/extensions/rules.scala
@@ -14,87 +14,21 @@
  */
 package org.apache.spark.sql.extensions
 
-import com.pingcap.tispark.statistics.StatisticsManager
-import com.pingcap.tispark.utils.ReflectionUtil._
-import com.pingcap.tispark.{MetaManager, TiDBRelation, TiTableReference}
-import org.apache.spark.sql.catalyst.TableIdentifier
-import org.apache.spark.sql.catalyst.analysis.{EliminateSubqueryAliases, UnresolvedRelation}
-import org.apache.spark.sql.catalyst.catalog.TiSessionCatalog
-import org.apache.spark.sql.catalyst.plans.logical.{InsertIntoTable, LogicalPlan}
+import com.pingcap.tispark.utils.ReflectionUtil
+import org.apache.spark.sql._
+import org.apache.spark.sql.catalyst.plans.logical.LogicalPlan
 import org.apache.spark.sql.catalyst.rules.Rule
-import org.apache.spark.sql.execution.command._
-import org.apache.spark.sql.execution.datasources.LogicalRelation
-import org.apache.spark.sql.{AnalysisException, _}
 
-case class TiResolutionRule(getOrCreateTiContext: SparkSession => TiContext)(
-    sparkSession: SparkSession)
-    extends Rule[LogicalPlan] {
-  protected lazy val meta: MetaManager = tiContext.meta
-  private lazy val autoLoad = tiContext.autoLoad
-  private lazy val tiCatalog = tiContext.tiCatalog
-  private lazy val tiSession = tiContext.tiSession
-  private lazy val sqlContext = tiContext.sqlContext
-  protected val tiContext: TiContext = getOrCreateTiContext(sparkSession)
-  protected val resolveTiDBRelation: TableIdentifier => LogicalPlan =
-    tableIdentifier => {
-      val dbName = getDatabaseFromIdentifier(tableIdentifier)
-      val tableName = tableIdentifier.table
-      val table = meta.getTable(dbName, tableName)
-      if (table.isEmpty) {
-        throw new AnalysisException(s"Table or view '$tableName' not found in database '$dbName'")
-      }
-      if (autoLoad) {
-        StatisticsManager.loadStatisticsInfo(table.get)
-      }
-      val sizeInBytes = StatisticsManager.estimateTableSize(table.get)
-      val tiDBRelation =
-        TiDBRelation(tiSession, TiTableReference(dbName, tableName, sizeInBytes), meta)(
-          sqlContext)
-      // Use SubqueryAlias so that projects and joins can correctly resolve
-      // UnresolvedAttributes in JoinConditions, Projects, Filters, etc.
-      newSubqueryAlias(tableName, LogicalRelation(tiDBRelation))
-    }
-
-  override def apply(plan: LogicalPlan): LogicalPlan =
-    plan transformUp resolveTiDBRelations
-
-  protected def resolveTiDBRelations: PartialFunction[LogicalPlan, LogicalPlan] = {
-    case i @ InsertIntoTable(UnresolvedRelation(tableIdentifier), _, _, _, _)
-        if tiCatalog
-          .catalogOf(tableIdentifier.database)
-          .exists(_.isInstanceOf[TiSessionCatalog]) =>
-      i.copy(table = EliminateSubqueryAliases(resolveTiDBRelation(tableIdentifier)))
-    case UnresolvedRelation(tableIdentifier)
-        if tiCatalog
-          .catalogOf(tableIdentifier.database)
-          .exists(_.isInstanceOf[TiSessionCatalog]) =>
-      resolveTiDBRelation(tableIdentifier)
+class TiResolutionRuleFactory(getOrCreateTiContext: SparkSession => TiContext)
+  extends (SparkSession => Rule[LogicalPlan]) {
+  override def apply(sparkSession: SparkSession): Rule[LogicalPlan] = {
+    ReflectionUtil.newTiResolutionRule(getOrCreateTiContext, sparkSession)
   }
-
-  private def getDatabaseFromIdentifier(tableIdentifier: TableIdentifier): String =
-    tableIdentifier.database.getOrElse(tiCatalog.getCurrentDatabase)
 }
 
-case class TiDDLRule(getOrCreateTiContext: SparkSession => TiContext)(sparkSession: SparkSession)
-    extends Rule[LogicalPlan] {
-  protected lazy val tiContext: TiContext = getOrCreateTiContext(sparkSession)
-
-  override def apply(plan: LogicalPlan): LogicalPlan =
-    plan transformUp {
-      // TODO: support other commands that may concern TiSpark catalog.
-      case sd: ShowDatabasesCommand =>
-        TiShowDatabasesCommand(tiContext, sd)
-      case sd: SetDatabaseCommand =>
-        TiSetDatabaseCommand(tiContext, sd)
-      case st: ShowTablesCommand =>
-        TiShowTablesCommand(tiContext, st)
-      case st: ShowColumnsCommand =>
-        TiShowColumnsCommand(tiContext, st)
-      case dt: DescribeTableCommand =>
-        TiDescribeTablesCommand(tiContext, dt)
-      case dc: DescribeColumnCommand =>
-        TiDescribeColumnCommand(tiContext, dc)
-      case ct: CreateTableLikeCommand =>
-        TiCreateTableLikeCommand(tiContext, ct)
-    }
+class TiDDLRuleFactory(getOrCreateTiContext: SparkSession => TiContext)
+  extends (SparkSession => Rule[LogicalPlan]) {
+  override def apply(sparkSession: SparkSession): Rule[LogicalPlan] = {
+    ReflectionUtil.newTiDDLRule(getOrCreateTiContext, sparkSession)
+  }
 }
