+++ spark-wrapper/spark-2.3/pom.xml
@@ -73,7 +73,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <scalaVersion>${scala.version}</scalaVersion>
+                    <scalaVersion>2.11.12</scalaVersion>
                 </configuration>
             </plugin>
             <plugin>
@@ -105,7 +105,7 @@
             <!-- Scala Format Plug-in -->
             <plugin>
                 <groupId>org.antipathy</groupId>
-                <artifactId>mvn-scalafmt_${scala.binary.version}</artifactId>
+                <artifactId>mvn-scalafmt_2.11</artifactId>
                 <version>1.0.3</version>
                 <configuration>
                     <skipSources>${scalafmt.skip}</skipSources>
