+++ spark-wrapper/spark-2.3/src/main/scala/org/apache/spark/sql/TiAggregationImpl.scala
@@ -117,11 +117,11 @@ object TiAggregationImpl {
           (sumAndAvgEliminated ++ extraSumsAndCounts ++ rewriteSums).distinct
         }
 
-        Some(
+        Some((
           groupingExpressions,
           rewrittenAggregateExpressions,
           rewrittenResultExpressions,
-          child)
+          child))
 
       case _ => Option.empty[ReturnType]
     }
